refactor: extract dataset ID logic to useDatasetId composable
